Introduce an "install-to-prefix"
Removes "intall-mod-path", "install-path" and "install-initramfs-path".

install-path is now $install-to-prefix/$bootdir
install-initramfs-path is now $install-to-prefix/$bootdir
install-mod-path is now $install-to-prefix

The default "install-to-prefix" is "/", but is changed with a warning if a user
cannot write to it.  Then, it is set to be in the "$TEMP/out" directory of that
user.

Other fixes and additions :
* The "--[no]-install" target is removed since anybody can install in a
  temporary prefix somewhere.  Every default genkernel run now installs the
  kernel, modules and initramfs.
* Always uses a kbuild_output in the persistent user's "cache" (and not /tmp)
  directory so multiple run of genkernel with "--no-mrproper" and "--no-clean"
  will reuse the compiled objects.
* Display size of main outputs (lib/, kernel, initramfs)
* Use the running kernel config automatically
* dist/ directory in user's cache



git-svn-id: svn://svn.berlios.de/genkernel/trunk@244 5ba8dd8a-5c08-0410-8eb7-bf276fad51ba